fix(hisparse): plan-buffer padding, async-backup removal, gather width
- produce_plan: refill the persistent plan hot_indices with -1 each
  step; the kernel skips CUDA-graph padding rows, which otherwise kept
  stale values and fed them to attention on replay (bounded but
  violating the documented -1 contract).
- Drop VLLM_HISPARSE_ASYNC_BACKUP: it aborts FULL_DECODE_ONLY capture
  (unjoined side-stream fork inside the captured region) and the target
  regime always runs graphs, so it was dead weight plus five no-op
  wait_for_pending_backup call sites.
- hisparse_gather_plan: launch 1024 threads (was 256) to match the
  swap-in kernel; the gather serves 3 of every 4 layers' misses, so
  per-row copy parallelism is the cold-row throughput limiter.
- host_to_device_ratio: parse as float instead of silently truncating
  fractional values.
- Drop validate_hisparse_decode_batch (no production callers).

+++ vllm/v1/core/kv_cache_utils.py
@@ -1353,7 +1353,7 @@ def _hisparse_host_pool_bytes(
         return None
     if config.host_pool_gib is not None:
         return int(config.host_pool_gib * 2**30)
-    return config.host_to_device_ratio * available_memory
+    return int(config.host_to_device_ratio * available_memory)
 
 
 def get_kv_cache_config_from_groups(
